+++ CHANGELOG.md
@@ -1,3 +1,25 @@
-## 0.1.0 ##
+## (Unreleased) ##
+
+* Fixes bug where CLI --no-color wasn't respected
+
+## 1.0.0 - 4/7/2013 ##
+
+* Allows the user to specify ignored properties and selectors
+* Better parse error messages
+
+## 0.2.1 - 3/28/2013 ##
+
+* Changes coloring to the selectors and declarations
+* Fixes bug where some duplicates weren't being combined #1
+
+## 0.2.0 - 3/24/2013 ##
+
+* Colorizes text output.
+* Supports scss/sass files.
+* Fixes newline output bug when there are no redundancies
+* Downloads remote css files if passed a URL
+* Fixes bug with double semicolons (blank attributes)
+
+## 0.1.0 - 3/21/2013 ##
 
 * Initial project release.